Created angular component for users to query events and subscribe to events
Created responsive angular component for users to query events and subscribe to events. Integrated it with the spring boot to get event data from the database. Created the event getter service in angular to handle requests about events. Waiting for event subscription api to connect the subscribe function to the backend.Also created sidebar component for the alumnus user