Change Header,view alumni
Created component to list all alumni, created a component to change the header by getting  image from the user and created a backend service to handle uploads and serve images to the client. Created a page that would allow users to view all events. Created a menubar component to replace the admin homepage for easier navigation.